Result count and list parameters for newspapers and departments now working

+++ src/main/scala/de/htwBerlin/ai/inews/core/ArticleList.scala
@@ -4,7 +4,7 @@ import spray.json.DefaultJsonProtocol._
 import spray.json.RootJsonFormat
 
 case class ArticleList(
-  resultCount: Int,
+  resultCount: Long,
   articles: Seq[Article]
 )
